@@ -1,23 +1,24 @@
 {
   "name": "report",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.21.1",
+    "axios": "^0.27.2",
     "classnames": "^2.3.1",
+    "net": "^1.0.2",
     "nodeclipse": "^0.17.2",
-    "primeflex": "^2.0.0",
-    "primeicons": "^4.1.0",
-    "primereact": "^6.6.0",
+    "primeflex": "^3.2.1",
+    "primeicons": "^5.0.0",
+    "primereact": "^8.2.0",
     "querystringify": "^2.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-intl": "^5.20.1",
-    "react-scripts": "^4.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-intl": "^6.0.5",
+    "react-scripts": "^5.0.1",
     "react-transition-group": "^4.4.2",
-    "sockjs-client": "^1.5.1",
+    "sockjs-client": "^1.6.1",
     "stompjs": "^2.3.3",
-    "typescript": "^4.3.2"
+    "typescript": "^4.7.4"
   },
   "scripts": {
     "start": "react-scripts start",
